+++ extensions/catppuccin/src/types.ts
@@ -4,6 +4,7 @@ export interface PortDetails {
   categories?: string[];
   icon?: string;
   color?: string;
+  alias?: string;
 }
 
 export interface UserStyleDetails {
@@ -25,8 +26,4 @@ export interface PortsYaml {
 
 export interface UserStylesYaml {
   userstyles: Record<string, UserStyleDetails>;
-}
-
-export interface Preferences {
-  gridSize: string;
 }